Fix sign-compare and unused warnings across samples
Address GCC -Wall and MSVC default warnings reported by the Windows
MSBuild CI:

- Switch most for-loop iterators to size_t when comparing against
  std::vector::size() (sign-compare / C4267 fix).
- Cast comparisons of int with .size()/.dim/.size[i] to int or
  size_t depending on context (mediapipe_iris, gpt-sovits,
  g2p_en_*, clip).
- Cast size_t -> (unsigned int) when passing to ailiaSetInputBlobData,
  ailiaGetBlobData, ailiaCreateDetector, ailiaTokenizerDecode, etc.
- Cast (int) on cv::Mat ctor / hsv_to_rgb / std::round to silence
  C4244 / C4267.
- Replace sprintf with snprintf in arcface and m2det (C4996).
- Add UTF-8 prefix u8"..." to bert_maskedlm Japanese literal (C4566).
- Replace exp() with expf() in bert_maskedlm softmax (C4244 double->
  float).
- Remove unused locals (whisper task/env_id, multilingual-e5/sentence_
  transformers status, t5_whisper_medical num_beams, retinaface
  flags/fontScale, mediapipe_iris mat_confidence) and the unused
  print_net helper in clap.
- Fix typo `cv:cvtColor` -> `cv::cvtColor` in mediapipe_iris (was
  parsed as a label).

After these changes, both `cmake .` and `cmake -DWITH_OPENCV=OFF .`
builds with `-Wall` produce zero warnings on GCC.

+++ background_removal/u2net/u2net.cpp
@@ -105,7 +105,7 @@ static void print_error(std::string arg)
 static void print_arch_error(std::string arg)
 {
     PRINT_ERR("u2net: error: argument -a/--arch: invalid choice: \'%s\' (choose from", arg.c_str());
-    for (int i = 0; i < MODEL_LISTS.size(); i++) {
+    for (size_t i = 0; i < MODEL_LISTS.size(); i++) {
         PRINT_ERR(" \'%s\'", MODEL_LISTS[i]);
         if (i < MODEL_LISTS.size() - 1) {
             PRINT_ERR(",");
@@ -119,7 +119,7 @@ static void print_arch_error(std::string arg)
 static void print_opset_error(std::string arg)
 {
     PRINT_ERR("u2net: error: argument -o/--opset: invalid choice: \'%s\' (choose from", arg.c_str());
-    for (int i = 0; i < OPSET_LISTS.size(); i++) {
+    for (size_t i = 0; i < OPSET_LISTS.size(); i++) {
         PRINT_ERR(" \'%s\'", OPSET_LISTS[i]);
         if (i < OPSET_LISTS.size() - 1) {
             PRINT_ERR(",");
@@ -170,7 +170,7 @@ static int argument_parser(int argc, char **argv)
             }
         }
         else if (arg[0] != '-') {
-            int j;
+            size_t j;
             switch (status) {
             case 1:
                 image_path = arg;
